+++ src/main/java/se/bjurr/prnfs/listener/UrlInvoker.java
@@ -7,9 +7,11 @@ import static java.lang.Boolean.TRUE;
 import static javax.ws.rs.core.HttpHeaders.AUTHORIZATION;
 import static javax.xml.bind.DatatypeConverter.printBase64Binary;
 
+import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.UnsupportedEncodingException;
+import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLConnection;
 
@@ -23,18 +25,32 @@ import com.google.common.io.Closeables;
 public class UrlInvoker {
  private static final Logger logger = LoggerFactory.getLogger(UrlInvoker.class);
 
- public void ivoke(String urlParam, Optional<String> user, Optional<String> password) {
+ public void ivoke(String urlParam, Optional<String> user, Optional<String> password, String method,
+   Optional<String> postContent) {
   InputStreamReader ir = null;
+  DataOutputStream wr = null;
   try {
    logger.info("Url: \"" + urlParam + "\"");
    final URL url = new URL(urlParam);
-   final URLConnection uc = url.openConnection();
+   final HttpURLConnection uc = (HttpURLConnection) url.openConnection();
+   uc.setRequestMethod(method);
    setAuthorization(uc, user, password);
+   uc.setDoOutput(true);
+   if (shouldPostContent(method, postContent)) {
+    logger.debug("POST >\n" + postContent.get());
+    uc.setDoInput(true);
+    uc.setRequestProperty("Content-Length", postContent.get().length() + "");
+    wr = new DataOutputStream(uc.getOutputStream());
+    wr.write(postContent.get().getBytes(UTF_8));
+   }
    ir = new InputStreamReader(uc.getInputStream(), UTF_8);
    logger.debug(on("\n").join(readLines(ir)));
   } catch (final Exception e) {
    try {
     Closeables.close(ir, TRUE);
+    if (wr != null) {
+     Closeables.close(wr, TRUE);
+    }
    } catch (final IOException e1) {
    }
    logger.error("", e);
@@ -51,6 +67,11 @@ public class UrlInvoker {
    uc.setRequestProperty(AUTHORIZATION, basicAuth);
   }
 
+ }
+
+ @VisibleForTesting
+ public static boolean shouldPostContent(String method, Optional<String> postContent) {
+  return method.equals("POST") && postContent.isPresent();
  }
 
  @VisibleForTesting